+++ docstrfmt/main.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 import asyncio
 import glob
+import itertools
 import logging
 import os
 import signal
@@ -38,7 +39,7 @@ from libcst import CSTTransformer, Expr
 from libcst.metadata import ParentNodeProvider, PositionProvider
 
 from . import __version__
-from .const import DEFAULT_EXCLUDE
+from .const import DEFAULT_EXCLUDE, SECTION_CHARS
 from .debug import dump_node
 from .docstrfmt import Manager
 from .exceptions import InvalidRstErrors
@@ -63,12 +64,14 @@ def _format_file(
     mode: Mode,
     docstring_trailing_line: bool,
     format_python_code_blocks: bool,
+    section_adornments: list[tuple[str, bool]] | None,
     raw_output: bool,
     lock: Lock,
 ):
     error_count = 0
     manager = Manager(
-        reporter, mode, docstring_trailing_line, format_python_code_blocks
+        reporter, mode, docstring_trailing_line, format_python_code_blocks,
+        section_adornments
     )
     if file.name == "-":
         raw_output = True
@@ -288,6 +291,25 @@ def _resolve_length(context: click.Context, _: click.Parameter, value: int | Non
     return value or pyproject_line_length
 
 
+def _validate_adornments(
+    context: click.Context, _: click.Parameter, value: str | None
+) -> list[tuple[str, bool]]:
+    if value is None:
+        return value
+
+    if len(value) != len(set(value)):
+        msg = "Section adornments must be unique"
+        raise click.BadParameter(msg)
+
+    if "|" in value:
+        with_overline, without_overline = value.split("|", 1)
+        return list(zip(with_overline, itertools.repeat(True))) + list(
+            zip(without_overline, itertools.repeat(False))
+        )
+
+    return list(zip(value, itertools.repeat(False)))
+
+
 async def _run_formatter(
     check: bool,
     file_type: str,
@@ -295,6 +317,7 @@ async def _run_formatter(
     include_txt: bool,
     docstring_trailing_line: bool,
     format_python_code_blocks: bool,
+    section_adornments: list[tuple[str, bool]] | None,
     mode: Mode,
     line_length: int,
     raw_output: bool,
@@ -322,6 +345,7 @@ async def _run_formatter(
                 mode,
                 docstring_trailing_line,
                 format_python_code_blocks,
+                section_adornments,
                 raw_output,
                 lock,
             )
@@ -465,14 +489,18 @@ class Visitor(CSTTransformer):
         )
 
     def leave_ClassDef(  # noqa: N802
-        self, original_node: ClassDef, updated_node: ClassDef  # noqa: ARG002
+        self,
+        original_node: ClassDef,
+        updated_node: ClassDef,  # noqa: ARG002
     ) -> ClassDef:
         """Remove the class name from the object name stack."""
         self._object_names.pop(-1)
         return updated_node
 
     def leave_FunctionDef(  # noqa: N802
-        self, original_node: FunctionDef, updated_node: FunctionDef  # noqa: ARG002
+        self,
+        original_node: FunctionDef,
+        updated_node: FunctionDef,  # noqa: ARG002
     ) -> FunctionDef:
         """Remove the function name from the object name stack."""
         self._object_names.pop(-1)
@@ -513,7 +541,7 @@ class Visitor(CSTTransformer):
             self.error_count += self.manager.error_count
             self.manager.error_count = 0
             object_display_name = (
-                f'{self._object_type} {".".join(self._object_names)!r}'
+                f"{self._object_type} {'.'.join(self._object_names)!r}"
             )
             single_line = len(output.splitlines()) == 1
             original_strip = original_node.evaluated_value.rstrip(" ")
@@ -645,6 +673,25 @@ class Visitor(CSTTransformer):
     callback=_resolve_length,
 )
 @click.option(
+    "-s",
+    "--section-adornments",
+    type=str,
+    default=None,
+    is_flag=False,
+    flag_value=SECTION_CHARS,
+    help=(
+        f"""Force pre-defined section adornments for part/chapter/section headers. If an
+        optional string is provided, it defines a sequence of adornments to use for each
+        individual section depth. The list must be composed of at least N **distinct**
+        characters for documents with N section depths.  Provide more if unsure.  If the
+        special character `|` (pipe) is used, then it defines sections (left portion)
+        that will have overlines besides underlines only (right portion). If this option
+        is not set, the default behaviour is to preserve existing adornments on your
+        document. An example set of adornments would be `{SECTION_CHARS}`."""
+    ),
+    callback=_validate_adornments,
+)
+@click.option(
     "-p",
     "--pyproject-config",
     "mode",
@@ -704,6 +751,7 @@ def main(  # noqa: PLR0912,PLR0915
     ignore_cache: bool,
     include_txt: bool,
     line_length: int,
+    section_adornments: list[tuple[str, bool]] | None,
     mode: Mode,
     quiet: bool,
     raw_input: str,
@@ -728,7 +776,8 @@ def main(  # noqa: PLR0912,PLR0915
     error_count = 0
     if raw_input:
         manager = Manager(
-            reporter, mode, docstring_trailing_line, format_python_code_blocks
+            reporter, mode, docstring_trailing_line, format_python_code_blocks,
+            section_adornments
         )
         file = "<raw_input>"
         check = False
@@ -763,6 +812,7 @@ def main(  # noqa: PLR0912,PLR0915
                 mode,
                 docstring_trailing_line,
                 format_python_code_blocks,
+                section_adornments,
                 raw_output,
                 None,
             )
@@ -794,6 +844,7 @@ def main(  # noqa: PLR0912,PLR0915
                     include_txt,
                     docstring_trailing_line,
                     format_python_code_blocks,
+                    section_adornments,
                     mode,
                     line_length,
                     raw_output,